+++ Model/Config/Provider.php
@@ -44,6 +44,7 @@ class Provider implements ConfigProviderInterface
         Config::METHOD_BANCONTACT,
         Config::METHOD_PREPAYMENT,
         Config::METHOD_APPLEPAY,
+        Config::METHOD_GOOGLEPAY,
     ];
 
     /**