+++ view/frontend/web/js/view/payment/method-renderer/paylater_invoice.js
@@ -10,6 +10,8 @@ define(
         'use strict';
 
         return Component.extend({
+            isThreatMetrixNeeded: true,
+
             defaults: {
                 template: 'Unzer_PAPI/payment/paylater_invoice',
                 fieldId: 'unzer-paylater-invoice-customer',
@@ -27,7 +29,7 @@ define(
             },
 
             _initializeCustomerFormForB2cCustomer: function (fieldId, errorFieldId, customer) {
-                threatMetrix.init(customer.threat_metrix_id);
+                threatMetrix.init(customer.threat_metrix_id, this);
 
                 this.resourceProvider.create({
                     containerId: fieldId + '-optin',